requirements.txt

Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
openai==0.28.0
requests==2.31.0
python-dotenv==1.0.1

# Production Requirements
gunicorn==21.2.0
whitenoise==6.6.0
